chore: extract error-tracking-throttler
