+++ app/lib/util/error-tracking.js
@@ -14,6 +14,8 @@ const log = require('../log')('app:error-tracking');
 
 const { rewriteFramesIntegration } = require('@sentry/integrations');
 
+const { createThrottler } = require('./error-tracking-throttler');
+
 const PRIVACY_PREFERENCES_CONFIG_KEY = 'editor.privacyPreferences';
 const EDITOR_ID_CONFIG_KEY = 'editor.id';
 const CRASH_REPORTS_CONFIG_KEY = 'ENABLE_CRASH_REPORTS';
@@ -21,13 +23,6 @@ const NON_EXISTENT_EDITOR_ID = 'NON_EXISTENT_EDITOR_ID';
 
 const SENTRY_DSN_FLAG = 'sentry-dsn';
 const DISABLE_REMOTE_INTERACTION_FLAG = 'disable-remote-interaction';
-
-const SIGNATURE_CAP = 5;
-const RENDER_LOOP_CAP = 1;
-
-const REACT_RENDER_LOOP_MESSAGES = [
-  /Maximum update depth exceeded/
-];
 
 let isActive = false;
 
@@ -177,51 +172,3 @@ function normalizeUrl(path) {
   const filename = path.replace(/^.*[\\\/]/, '');
   return 'file:///build/' + filename;
 }
-
-/**
- * Build a Sentry `beforeSend` callback that drops non-actionable events and
- * caps repeated reports per error signature within a session.
- */
-function createThrottler() {
-  const counts = new Map();
-
-  return function beforeSend(event) {
-    const value = getExceptionValue(event);
-
-    const isRenderLoop = REACT_RENDER_LOOP_MESSAGES.some(re => re.test(value));
-    const cap = isRenderLoop ? RENDER_LOOP_CAP : SIGNATURE_CAP;
-
-    const signature = getSignature(event);
-    const count = counts.get(signature) || 0;
-
-    if (count >= cap) {
-      return null;
-    }
-
-    counts.set(signature, count + 1);
-    return event;
-  };
-}
-
-function getExceptionValue(event) {
-  const values = event && event.exception && event.exception.values;
-
-  if (values && values.length) {
-    return (values[0] && values[0].value) || '';
-  }
-
-  return (event && event.message) || '';
-}
-
-function getSignature(event) {
-  const values = event && event.exception && event.exception.values;
-
-  if (values && values.length) {
-    const v = values[0] || {};
-    return (v.type || 'Error') + ':' + (v.value || '');
-  }
-
-  return (event && event.message) || 'unknown';
-}
-
-module.exports.createThrottler = createThrottler;
